requirements/development.txt
-r ./test.txt

black>=24.1.1
ruff>=0.1.14
isort>=5.13.2
djlint>=1.34.1
bandit>=1.7.7
django-types>=0.19.1
djangorestframework-types>=0.8.0
